Add data retrieval APIs for UART and CAN tasks
Introduced functions to retrieve received UART and CAN data in both C and C++ interfaces. Updated timer configuration for TIM1 to use a prescaler of 72-1 and period of 1000-1, and refactored the timer callback logic for improved LED toggling. Fixed DMA memory alignment for USART2 RX to byte alignment. Updated related struct definitions and main loop to utilize the new data retrieval functions.

+++ Core/Inc/tim.h
@@ -42,7 +42,7 @@ void MX_TIM1_Init(void);
 
 /* USER CODE BEGIN Prototypes */
 extern volatile uint32_t tick;
-extern uint8_t tim_count_led;
+extern uint16_t tim_count_led;
 /* USER CODE END Prototypes */
 
 #ifdef __cplusplus
